Sonar cube dedup issue
