+++ tests/integration/test_metrics_service_command.py
@@ -18,6 +18,7 @@ from django.utils import timezone
 
 from apps.tasks.models import Task
 from tests.test_utils import get_test_password
+from tests.unit.core.test_metrics_service_helpers import create_mock_processes_with_exit, get_default_config
 
 User = get_user_model()
 
@@ -335,47 +336,12 @@ class TestMetricsServiceFullIntegration(TransactionTestCase):
         self.manage_py = Path(__file__).parent.parent.parent / "manage.py"
         assert self.manage_py.exists(), f"manage.py not found at {self.manage_py}"
 
-    def _create_mock_processes_with_exit(self):
-        """Create mock processes where Django process exits after first check."""
-        poll_call_count = {"django": 0}
-
-        def django_poll_side_effect():
-            """Simulate Django process that exits after first monitoring check."""
-            poll_call_count["django"] += 1
-            if poll_call_count["django"] == 1:
-                return None  # First check: still running
-            return 0  # Exited
-
-        django_process = MagicMock()
-        django_process.poll.side_effect = django_poll_side_effect
-        django_process.returncode = 0
-        django_process.terminate.return_value = None
-        django_process.kill.return_value = None
-
-        other_process = MagicMock()
-        other_process.poll.return_value = None
-        other_process.terminate.return_value = None
-        other_process.kill.return_value = None
-
-        return [django_process, other_process, other_process]
-
-    def _get_default_config(self):
-        """Get default test configuration."""
-        return {
-            "host": "127.0.0.1",
-            "port": "8000",
-            "workers": 4,
-            "log_level": "INFO",
-            "timeout": 3600,
-            "max_tasks": 100,
-        }
-
     def _setup_mocks_for_start_services(self, mock_exists, mock_exit, mock_sleep, mock_popen):
         """Set up common mocks for tests that call _start_services."""
         mock_exists.return_value = True
         mock_exit.side_effect = SystemExit
         mock_sleep.return_value = None
-        mock_popen.side_effect = self._create_mock_processes_with_exit()
+        mock_popen.side_effect = create_mock_processes_with_exit()
 
     def _create_command_instance(self):
         """Create and configure a Command instance for testing."""
@@ -438,7 +404,7 @@ class TestMetricsServiceFullIntegration(TransactionTestCase):
         """Test that signal handlers are properly configured in _start_services_simple."""
         command = self._create_command_instance()
         self._setup_mocks_for_start_services(mock_exists, mock_exit, mock_sleep, mock_popen)
-        config = self._get_default_config()
+        config = get_default_config()
 
         # Signal handlers are set up inside _start_services_simple
         with pytest.raises(SystemExit):
@@ -521,7 +487,7 @@ class TestMetricsServiceFullIntegration(TransactionTestCase):
         command = self._create_command_instance()
         manage_py = Path(__file__).parent.parent.parent / "manage.py"
         self._setup_mocks_for_start_services(mock_exists, mock_exit, mock_sleep, mock_popen)
-        config = self._get_default_config()
+        config = get_default_config()
 
         # Start services will build the dispatcher command internally
         # We'll verify it by checking the Popen calls
